feat: enhance security and authentication features
- Added email-validator and pyjwt to requirements for improved email validation and JWT handling.
- Introduced bcrypt for password hashing in security configuration.
- Implemented a new method in AgentProfileLoader to retrieve all agent configurations.
- Updated setup.sh with security-related environment variables for JWT and session management.
- Created a comprehensive test suite for authentication services, including password and JWT operations.
- Developed user model tests to ensure proper functionality and validation.
- Established database operations tests for user creation and retrieval.
- Added configuration for test database setup and teardown in pytest fixtures.

+++ data/db/models/transcript.py
@@ -1,7 +1,7 @@
 """
 Transcript database model.
 """
-from sqlalchemy import Column, String, Text, DateTime, Float, ForeignKey, Boolean
+from sqlalchemy import Column, String, Text, DateTime, Float, Integer, ForeignKey, Boolean
 from sqlalchemy.orm import relationship
 from data.db.base import BaseModel
 
@@ -11,7 +11,7 @@ class Transcript(BaseModel):
   __tablename__ = "transcripts"
 
   # Link to call log
-  call_log_id = Column(Float, ForeignKey(
+  call_log_id = Column(Integer, ForeignKey(
       "call_logs.id"), nullable=False, index=True)
   call_log = relationship("CallLog", back_populates="transcripts")
